Do not send empty TXT records
RFC6763 6.1 states that empty TXT records are not allowed.
Also every DNS-SD service must have a TXT record, even if
it is empty.

We didn't check that before, which had the effect that
service instance names were not resolved on macOS and iOS
using the Discovery.app.

This fix sets a minimum of one empty TXT record.